Fixes for porting on VC++7.1. std::pow(double,double) instead of (int,double)
git-svn-id: svn+ssh://svn.cern.ch/reps/geant4/trunk/geant4@22361 82ad68bf-922b-4865-b672-c5fc3b8c97f6

+++ source/processes/hadronic/models/coherent_elastic/include/G4DiffElasticHadrNucleus.hh
@@ -1,3 +1,29 @@
+//
+// ********************************************************************
+// * DISCLAIMER                                                       *
+// *                                                                  *
+// * The following disclaimer summarizes all the specific disclaimers *
+// * of contributors to this software. The specific disclaimers,which *
+// * govern, are listed with their locations in:                      *
+// *   http://cern.ch/geant4/license                                  *
+// *                                                                  *
+// * Neither the authors of this software system, nor their employing *
+// * institutes,nor the agencies providing financial support for this *
+// * work  make  any representation or  warranty, express or implied, *
+// * regarding  this  software system or assume any liability for its *
+// * use.                                                             *
+// *                                                                  *
+// * This  code  implementation is the  intellectual property  of the *
+// * GEANT4 collaboration.                                            *
+// * By copying,  distributing  or modifying the Program (or any work *
+// * based  on  the Program)  you indicate  your  acceptance of  this *
+// * statement, and all its terms.                                    *
+// ********************************************************************
+//
+//
+// $Id: G4DiffElasticHadrNucleus.hh,v 1.10 2005-06-10 13:23:42 gcosmo Exp $
+// GEANT4 tag $Name: not supported by cvs2svn $
+//
 #ifndef G4DiffElasticHadrNucleus_h
 #define G4DiffElasticHadrNucleus_h 1
 
@@ -128,7 +154,8 @@ class G4DiffElasticHadrNucleus: public //G4HadronValues
 
           if(N < 100) for(G4int M = 1; M<=N; M++)  Res = Res*M;         
 
-          if(N >= 100)  Res = 2.50662827*std::exp(static_cast<double>(-N-1))*std::pow(N+1,N+0.5)*
+          if(N >= 100)  Res = 2.50662827*std::exp(static_cast<double>(-N-1))*
+                         std::pow(static_cast<double>(N+1),N+0.5)*
                          (1+1/12/(N+1)+1/288/(N+1)/(N+1)-
                          139/51840/(N+1)/(N+1)/(N+1)-
                          571/2488320/(N+1)/(N+1)/(N+1)/(N+1));
